fixed : 경로 수정 및 db upload 삭제
