+++ src/app/api/img-upload/route.ts
@@ -1,0 +1,32 @@
+// // // app/api/documents/route.ts
+import { NextRequest, NextResponse } from "next/server";
+import { v4 as uuid } from "uuid";
+import { uploadImageToS3 } from "@/app/utils/uploadImageToS3";
+
+export const dynamic = "force-dynamic";
+export async function POST(request: NextRequest, response: NextResponse) {
+  try {
+    const formData = await request.formData();
+    const file = formData.get("file") as Blob | null;
+    if (!file) {
+      return NextResponse.json(
+        { error: "File blob is required." },
+        { status: 400 }
+      );
+    }
+
+    const mimeType = file.type;
+    const fileExtension = mimeType.split("/")[1];
+
+    const buffer = Buffer.from(await file.arrayBuffer());
+    const fileName = await uploadImageToS3(
+      buffer,
+      uuid() + "." + fileExtension
+    );
+
+    return NextResponse.json({ success: true, fileName });
+  } catch (error) {
+    console.error("Error uploading image:", error);
+    NextResponse.json({ message: "Error uploading image" });
+  }
+}
